demo/example_disbi_project/requirements.txt
django>=1.9,<2
django-import-export==0.5.0
numpy>=1.11.0
matplotlib>=1.5.1,<2
more-itertools
psycopg2-binary
django-admin-view-permission
gunicorn